Visual changes.
Hello Volmer,

I made some changes to the visual styles and functionalities of bootsy.

The visual changes were done for a better user experience so it's
easier for clients to use.

The visual styles include:
- z-index fix to allow the contextual drop-down menus to properly
appear over the footer and modal if it includes a longer list
- a default message if there are no uploaded images.
- a nicer looking file input

The functional changes include:
- auto submission on file input change

All tests are passing, however, I needed to change a few of them to
allow for the changes I made.
I needed to remove the step 'I press "Load"' as that is no longer
required.

Let me know if you have any questions.

ac.